Changed directory structure anticipating python package development. Fixed Makefile.
