--- OB151212coords.txt
@@ -1,1 +1,0 @@
-17:57:05 -30:22:59
