+++ VBBinaryLensing/data/OB151212coords.txt
@@ -1,0 +1,1 @@
+17:57:05 -30:22:59
