+++ VBBinaryLensingLibrary.egg-info/SOURCES.txt
@@ -1,0 +1,10 @@
+README.md
+setup.py
+VBBinaryLensingLibrary.egg-info/PKG-INFO
+VBBinaryLensingLibrary.egg-info/SOURCES.txt
+VBBinaryLensingLibrary.egg-info/dependency_links.txt
+VBBinaryLensingLibrary.egg-info/not-zip-safe
+VBBinaryLensingLibrary.egg-info/requires.txt
+VBBinaryLensingLibrary.egg-info/top_level.txt
+src/VBBinaryLensingLibrary.cpp
+src/bindings.cpp
