+++ pyVBB.egg-info/requires.txt
@@ -1,0 +1,1 @@
+pybind11>=2.2
